fix(scripts): Error handle pacscript functions (#298)
* fix(scripts): Add guards to {prepare,build,install}

* Add `error_log` to guards

* Add guards to removescript

* Fix error number

* Add guard to postinst for deb packages

* Fix spacing on error_log

Former-commit-id: 737bc42999e41e1368adb11bb923233bb6038459

+++ misc/scripts/install-local.sh
@@ -146,17 +146,17 @@ function makeVirtualDeb {
 	fancy_message info "Creating dummy package"
 	sudo mkdir -p "$SRCDIR/$name-pacstall/DEBIAN"
 	printf "Package: %s\n" "$name" | sudo tee "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
-	
+
 	if [[ $version =~ ^[0-9] ]]; then
 		printf "Version: %s-1\n" "$version" | sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
 	else
 		printf "Version: 0%s-1\n" "$version" | sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
 	fi
-	
+
 	if [[ -n $depends ]]; then
 		printf "Depends: %s\n" "${depends//' '/' , '}"| sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
 	fi
-	
+
 	if [[ -n $optdepends ]]; then
 		fancy_message info "$name has optional dependencies that can enhance its functionalities"
 		echo "Optional dependencies:"
@@ -165,12 +165,12 @@ function makeVirtualDeb {
 		if [[ $answer -eq 1 ]]; then
 			optinstall='--install-suggests'
 		fi
-		
+
 		printf "Suggests:" |sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
 		printf " %s\n" "${optdepends[@]}" | awk -F': ' '{print $1}' | tr '\n' ',' | head -c -2 | sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
 		printf "\n" | sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
 	fi
-	
+
 	printf "Architecture: all
 Essential: no
 Section: Pacstall
@@ -184,7 +184,7 @@ Replace: ${replace//' '/', '}" | sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/cont
 	printf "Provides: ${gives:-$name}
 Maintainer: ${maintainer:-Pacstall <pacstall@pm.me>}
 Description: This is a symbolic package used by pacstall, may be removed with apt or dpkg. $description\n" | sudo tee -a "$SRCDIR/$name-pacstall/DEBIAN/control" > /dev/null
-	
+
 	echo '#!/bin/bash
 if [[ PACSTALL_REMOVE != "true" ]]; then
 	source /var/cache/pacstall/'"$name"'/'"$version"'/'"$name"'.pacscript 2>&1 /dev/null
@@ -289,7 +289,7 @@ if [[ -n "$pacdeps" ]]; then
 		fancy_message info "Installing $i"
 		# If /tmp/pacstall-pacdeps-"$i" is available, it will trigger the logger to log it as a dependency
 		sudo touch /tmp/pacstall-pacdeps-"$i"
-		
+
 		if ! pacstall -P -I "$i"; then
 			fancy_message error "Failed to install pacstall dependencies"
 			error_log 8 "install $PACKAGE"
@@ -366,7 +366,7 @@ function hashcheck() {
 		if [[ "$url" != *".deb" ]]; then
 			sudo dpkg -r "$name" > /dev/null
 		fi
-		
+
 		fancy_message info "Cleaning up"
 		cleanup
 		return 1
@@ -423,7 +423,11 @@ case "$url" in
 		if sudo apt install -y -f ./"${url##*/}" 2>/dev/null; then
 			log
 			if type -t postinst > /dev/null 2>&1; then
-				postinst
+				if ! postinst; then
+					error_log 5 "postinst hook"
+					fancy_message error "Could not run postinst hook successfully"
+					exit 1
+				fi
 			fi
 
 			fancy_message info "Storing pacscript"
@@ -460,8 +464,8 @@ case "$url" in
 esac
 
 if [[ -n $patch ]]; then
-		fancy_message info "Downloading patches"
-		mkdir -p PACSTALL_patchesdir
+	fancy_message info "Downloading patches"
+	mkdir -p PACSTALL_patchesdir
 	for i in "${patch[@]}"; do
 		wget -q "$i" -P PACSTALL_patchesdir &
 	done
@@ -472,7 +476,10 @@ fi
 export pkgdir="/usr/src/pacstall/$name"
 
 fancy_message info "Preparing"
-prepare
+if ! prepare; then
+	fancy_message error "Could not prepare $PACKAGE properly"
+	exit 1
+fi
 
 # Check if build function doesn't exist
 if ! type -t build > /dev/null 2>&1; then
@@ -485,13 +492,21 @@ if ! type -t build > /dev/null 2>&1; then
 fi
 
 fancy_message info "Building"
-build
+if ! build; then
+	error_log 5 "build $PACKAGE"
+	fancy_message error "Could not properly build $PACKAGE"
+	exit 1
+fi
 
 # Trap so that we can clean up (hopefully without messing up anything)
 trap - SIGINT
 
 fancy_message info "Installing"
-install
+if ! install; then
+	error_log 14 "install $PACKAGE"
+	fancy_message error "Could not install $PACKAGE properly"
+	exit 1
+fi
 
 if [[ $REMOVE_DEPENDS = y ]]; then
 	sudo apt-get remove $build_depends
@@ -531,7 +546,11 @@ fi
 # `hash -r` updates PATH database
 hash -r
 if type -t postinst > /dev/null 2>&1; then
-	postinst
+	if ! postinst; then
+		error_log 5 "postinst hook"
+		fancy_message error "Could not run postinst hook successfully"
+		exit 1
+	fi
 fi
 
 fancy_message info "Storing pacscript"
